split sidebar in components
